% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{roffice_deps}
\alias{roffice_deps}
\title{Dependencies for "office-ui-fabric-react"}
\usage{
roffice_deps()
}
\value{
htmltools::htmlDependency
}
\description{
Dependencies for "office-ui-fabric-react"
}
